+++ src/controller/server_side.py
@@ -4,8 +4,8 @@ import prettytable
 import sys
 sys.path.insert(0, '..')
 
-from data import data
-from data.data import BUFFER_SIZE
+from src.data import data
+from src.data.data import BUFFER_SIZE
 
 
 class server_side:
