+++ src/data/data.py
@@ -38,7 +38,7 @@ TIMEOUT = 12
 # paths
 CLIENT_NAME = "alvd"
 WIN_STARTUP_PATH = f"C:\\Users\\{os.getlogin()}\\AppData\\Roaming\\Microsoft\\Windows\\Start Menu\\Programs\\Startup"
-DEFAULT_ICON = "data\\images\\logoicon.ico"
+DEFAULT_ICON = "src\\data\\images\\logoicon.ico"
